4002 refactor file and image uploader
