+++ libs/admin/src/lib/admin-panel/components/organization/forms/organization-form/organization-form.component.html
@@ -34,9 +34,9 @@
       <h2>Company Logo</h2>
       <h5>Upload company Logo</h5>
 
-      <image-new-uploader ratio="square" [form]="form.get('logo')" [meta]="['orgs', 'logo', orgId]">
+      <image-uploader ratio="square" [form]="form.get('logo')" [meta]="['orgs', 'logo', orgId]">
         <h3 title>Logo</h3>
-      </image-new-uploader>
+      </image-uploader>
     </article>
   </section>
