+++ libs/campaign/src/lib/form/budget/budget.component.html
@@ -79,9 +79,9 @@
         <br />
         This upload is optional, but this information might be very helpful to your potential investors.
       </p>
-      <file-new-uploader [form]="form.get('files').get('budget').get('storagePath')" [meta]="['campaigns', 'budget', movieId]" accept="pdf">
+      <file-uploader [form]="form.get('files').get('budget').get('storagePath')" [meta]="['campaigns', 'budget', movieId]" accept="pdf">
         <h3 title>Upload File</h3>
-      </file-new-uploader>
+      </file-uploader>
       <span class="mat-caption">Upload a complete document.</span>
       <span class="mat-caption">Please note that this document will be available for download on the marketplace.</span>
     </section>
