+++ libs/campaign/src/lib/form/fundings/fundings.component.html
@@ -80,9 +80,9 @@
         <br />
         This upload is optional, but this information might be very helpful to your potential investors.
       </p>
-      <file-new-uploader [form]="form.get('files').get('financingPlan').get('storagePath')" [meta]="['campaigns', 'financingPlan', movieId]" accept="pdf">
+      <file-uploader [form]="form.get('files').get('financingPlan').get('storagePath')" [meta]="['campaigns', 'financingPlan', movieId]" accept="pdf">
         <h3 title>Upload File</h3>
-      </file-new-uploader>
+      </file-uploader>
       <span class="mat-caption">Please note that this document will be available for download on the marketplace.</span>
     </section>
   </ng-container>
