+++ libs/campaign/src/lib/form/profits/profits.module.ts
@@ -6,7 +6,7 @@ import { FlexLayoutModule } from '@angular/flex-layout';
 
 import { CampaignFormProfitsComponent } from './profits.component';
 
-import { FileNewUploaderModule } from '@blockframes/media/file/file-new-uploader/file-uploader.module';
+import { FileUploaderModule } from '@blockframes/media/file/file-uploader/file-uploader.module';
 import { TunnelPageModule } from '@blockframes/ui/tunnel';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
@@ -21,7 +21,7 @@ import { MatDividerModule } from '@angular/material/divider';
     ReactiveFormsModule,
     FlexLayoutModule,
     TunnelPageModule,
-    FileNewUploaderModule,
+    FileUploaderModule,
     MatFormFieldModule,
     MatInputModule,
     MatIconModule,
