+++ libs/media/src/lib/file/explorer/uploader-dialog/uploader-dialog.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { FlexLayoutModule } from '@angular/flex-layout';
 
-import { FileNewUploaderModule } from '@blockframes/media/file/file-new-uploader/file-uploader.module';
+import { FileUploaderModule } from '@blockframes/media/file/file-uploader/file-uploader.module';
 import { FormDisplayNameModule } from '@blockframes/ui/form/display-name/display-name.module';
 
 import { FileExplorerUploaderDialogComponent } from './uploader-dialog.component';
@@ -24,7 +24,7 @@ import { MatSelectModule } from '@angular/material/select';
     ReactiveFormsModule,
     FlexLayoutModule,
 
-    FileNewUploaderModule,
+    FileUploaderModule,
     FormDisplayNameModule,
 
     // Material
