--- libs/media/src/lib/image/uploader-new/uploader.component.html
@@ -1,54 +1,0 @@
-<input #fileUploader multiple hidden type="file" (change)="filesSelected($event.target.files)" />
-<section [ngSwitch]="step$ | async">
-
-  <!-- Drop -->
-  <article *ngSwitchCase="'drop'" class="drop-zone">
-    <button mat-icon-button aria-label="file upload button" (click)="fileUploader.click()">
-      <img class="add-files-icon" asset="add_files.webp" alt="upload image">
-    </button>
-    <ng-content select="[title]"></ng-content>
-    <p class="mat-body-2">{{ cropDimensions }}</p>
-    <p class="mat-caption">Accepted Formats: .jpg, .png</p>
-    <p class="mat-caption">Drag and drop a file here or <a (click)="fileUploader.click()">browse</a></p>
-  </article>
-
-  <!-- Hovering -->
-  <article *ngSwitchCase="'hovering'" class="drop-zone">
-    <img asset="add_files.webp" alt="upload image" class="picture-hover">
-    <h3>Drop your file here</h3>
-  </article>
-
-  <!-- Crop -->
-  <article *ngSwitchCase="'crop'">
-    <image-cropper [imageFile]="file" [aspectRatio]="cropRatio" format="webp" [containWithinAspectRatio]="false"
-      [resizeToWidth]="setWidth || parentWidth" (imageCropped)="imageCropped($event)"></image-cropper>
-    <div class="buttons" fxLayout="row" fxLayoutAlign="space-evenly center">
-      <button mat-mini-fab color="primary">
-        <mat-icon (click)="cropIt()" svgIcon="check"></mat-icon>
-      </button>
-      <button mat-mini-fab color="warn">
-        <mat-icon (click)="nextStep('drop')" svgIcon="close">cancel</mat-icon>
-      </button>
-    </div>
-  </article>
-
-  <!-- Show -->
-  <article class="show" *ngSwitchCase="'show'">
-    <img class="croppedImage" [src]="croppedImage || (previewUrl$ | async)" />
-    <div class="buttons" fxLayout="row" fxLayoutAlign="space-evenly center">
-      <ng-container *ngIf="useFileUploader">
-        <button mat-mini-fab color="primary">
-          <mat-icon (click)="fileUploader.click()" svgIcon="image" matTooltip="Select new picture"></mat-icon>
-        </button>
-      </ng-container>
-      <button mat-mini-fab color="accent" *ngIf="(previewUrl$ | async) as url">
-        <mat-icon [cdkCopyToClipboard]="url" svgIcon="archive" matTooltip="Copy image url"></mat-icon>
-      </button>
-      <ng-container *ngIf="useDelete">
-        <button mat-mini-fab color="warn">
-          <mat-icon (click)="delete()" svgIcon="delete" matTooltip="Delete current picture"></mat-icon>
-        </button>
-      </ng-container>
-    </div>
-  </article>
-</section>
