--- libs/media/src/lib/image/uploader-new/uploader.component.scss
@@ -1,54 +1,0 @@
-section {
-  width: 100%;
-  article {
-    text-align: center;
-    border: solid 1px var(--foreground-divider);
-    border-radius: 10px;
-    background: rgba(var(--primary), 0.3);
-    image-cropper {
-      width: inherit;
-      padding: 0;
-      margin: 5px;
-    }
-  }
-  article.drop-zone {
-    border: dashed 1px var(--primary);
-  }
-}
-
-a {
-  color: var(--primary);
-  cursor: pointer;
-}
-
-.mat-icon-button {
-  width: 48px;
-  height: 100%;
-  margin-bottom: 8px;
-  .add-files-icon {
-    width: 100%;
-  }
-}
-
-
-.mat-button {
-  padding: 0;
-  margin: 0;
-}
-
-.croppedImage {
-  margin-top: 5px;
-  max-width: 500px;
-  max-height: 1000px;
-}
-
-.buttons {
-  mat-icon {
-    cursor: pointer;
-  }
-  padding: 20px;
-}
-
-caption {
-  width: 100%;
-}
