+++ libs/media/src/lib/image/uploader/uploader.component.ts
@@ -1,12 +1,16 @@
 import { Component, Input, ChangeDetectionStrategy, OnInit, HostListener, ElementRef, ViewChild } from '@angular/core';
 import { ImageCroppedEvent } from 'ngx-image-cropper';
 import { BehaviorSubject } from 'rxjs';
-import { HostedMediaForm } from '@blockframes/media/form/media.form';
-import { MediaService } from '@blockframes/media/+state/media.service';
-import { ImageParameters } from '@blockframes/media/image/directives/imgix-helpers';
-import { getStoragePath, sanitizeFileName, Privacy, getMimeType } from '@blockframes/utils/file-sanitizer';
+import { MediaService } from '../../+state/media.service';
+import { ImageParameters } from '../../image/directives/imgix-helpers';
+import { sanitizeFileName, getMimeType } from '@blockframes/utils/file-sanitizer';
 import { SafeUrl, DomSanitizer } from '@angular/platform-browser';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { FormControl } from '@angular/forms';
+import { FileMetaData } from '../../+state/media.model';
+import { CollectionHoldingFile, FileLabel, getFileMetadata, getFileStoragePath } from '../../+state/static-files';
+import { FileUploaderService } from '../../+state/file-uploader.service';
+import { StorageFile } from '../../+state/media.firestore';
 
 type CropStep = 'drop' | 'crop' | 'hovering' | 'show';
 
@@ -34,7 +38,7 @@ function b64toBlob(data: string) {
   return new Blob([ab], { type });
 }
 @Component({
-  selector: 'image-uploader',
+  selector: '[meta] image-uploader',
   templateUrl: './uploader.component.html',
   styleUrls: ['./uploader.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
@@ -45,7 +49,7 @@ export class ImageUploaderComponent implements OnInit {
   // Private Variables //
   //////////////////////
 
-  private ref: string;
+  private storagePath: string;
   private step: BehaviorSubject<CropStep> = new BehaviorSubject('drop');
   private parameters: ImageParameters = {
     auto: 'compress,format',
@@ -65,6 +69,10 @@ export class ImageUploaderComponent implements OnInit {
   parentWidth: number;
   prev: CropStep;
   previewUrl$ = new BehaviorSubject<string | SafeUrl>('');
+
+  metadata: FileMetaData;
+  fileName: string;
+
 
   /////////////
   // Inputs //
@@ -90,13 +98,21 @@ export class ImageUploaderComponent implements OnInit {
         throw new Error('Unknown ratio');
     }
   }
-  @Input() form?: HostedMediaForm;
+  @Input() form: FormControl;
+
+  @Input() set meta(value: [CollectionHoldingFile, FileLabel, string] | [CollectionHoldingFile, FileLabel, string, number]) {
+    const [ collection, label, docId, index ] = value;
+    this.storagePath = getFileStoragePath(collection, label);
+    this.metadata = getFileMetadata(collection, label, docId, index);
+  }
+
+  @Input() index: number;
+
   @Input() setWidth?: number;
-  @Input() storagePath: string;
   /** Disable fileUploader & delete buttons in 'show' step */
   @Input() useFileUploader?= true;
   @Input() useDelete?= true;
-  @Input() filePrivacy: Privacy = 'public';
+
   @Input() types: string[] = ['image/jpeg', 'image/png'];
   @Input() accept: string[] = ['.jpg', '.png'];
 
@@ -105,17 +121,18 @@ export class ImageUploaderComponent implements OnInit {
   constructor(
     private mediaService: MediaService,
     private sanitizer: DomSanitizer,
-    private snackBar: MatSnackBar
+    private snackBar: MatSnackBar,
+    private uploaderService: FileUploaderService,
   ) { }
 
   ngOnInit() {
-    if (!!this.form.blobOrFile.value) {
-      const blobUrl = URL.createObjectURL(this.form.blobOrFile.value);
+    const retrieved = this.uploaderService.retrieveFromQueue(this.storagePath, this.index);
+    if (!!retrieved) {
+      const blobUrl = URL.createObjectURL(retrieved);
       const previewUrl = this.sanitizer.bypassSecurityTrustUrl(blobUrl);
       this.previewUrl$.next(previewUrl);
       this.nextStep('show');
-    } else if (!!this.form.oldRef?.value) {
-      this.ref = this.form.oldRef.value;
+    } else {
       this.goToShow();
     }
   }
@@ -143,7 +160,8 @@ export class ImageUploaderComponent implements OnInit {
   }
 
   private resetState() {
-    if (!!this.form.blobOrFile.value || (!!this.form.ref?.value)) {
+    const retrieved = this.uploaderService.retrieveFromQueue(this.storagePath, this.index);
+    if (!!retrieved) {
       this.nextStep('show');
     } else {
       this.nextStep('drop');
@@ -155,7 +173,13 @@ export class ImageUploaderComponent implements OnInit {
   ///////////
 
   async goToShow() {
-    this.previewUrl$.next(await this.getDownloadUrl(this.ref));
+    this.previewUrl$.next(await this.getDownloadUrl({
+      privacy: this.metadata.privacy,
+      collection: this.metadata.collection,
+      docId: this.metadata.docId,
+      field: this.metadata.field,
+      storagePath: this.storagePath
+    }));
     this.nextStep('show');
   }
 
@@ -178,7 +202,8 @@ export class ImageUploaderComponent implements OnInit {
       this.delete();
     } else {
       this.nextStep('crop');
-      this.form.patchValue({ cropped: false });
+      // TODO keep track of crop state
+      // this.form.patchValue({ cropped: false });
       this.fileUploader.nativeElement.value = null;
     }
 
@@ -204,13 +229,14 @@ export class ImageUploaderComponent implements OnInit {
       // and also postfix file name with a small random id allow the same image to be cropped several time without collision
       const fileName = sanitizeFileName(this.file.name.replace(/(\.[\w\d_-]+)$/i, `-${Math.random().toString(36).substr(2)}.webp`));
 
-      this.form.patchValue({
-        ref: getStoragePath(this.storagePath, this.filePrivacy),
-        blobOrFile: blob,
+      this.uploaderService.addToQueue(this.storagePath, {
         fileName: fileName,
-        cropped: true
-      })
-      this.form.markAsDirty();
+        file: blob,
+        metadata: this.metadata
+      });
+
+      // TODO keep track of crop state
+      // cropped state = true
 
     } catch (err) {
       console.error(err);
@@ -222,8 +248,9 @@ export class ImageUploaderComponent implements OnInit {
       this.croppedImage = '';
     }
 
-    this.form.patchValue({ ref: '', blobOrFile: undefined });
-    this.form.markAsDirty();
+    this.uploaderService.removeFromQueue(this.storagePath, this.fileName)
+    this.form.setValue('');
+
     this.fileUploader.nativeElement.value = null;
 
     this.nextStep('drop');
@@ -238,7 +265,7 @@ export class ImageUploaderComponent implements OnInit {
    * Returns a promise with the download url of an image based on its reference.
    * If media is protected, this will also try to fetch a security token.
    * */
-  private getDownloadUrl(ref: string): Promise<string> {
-    return this.mediaService.generateImgIxUrl(ref, this.parameters);
+  private getDownloadUrl(file: StorageFile): Promise<string> {
+    return this.mediaService.generateImgIxUrl(file, this.parameters);
   }
 }
