+++ libs/movie/src/lib/movie/form/main/main.component.html
@@ -96,9 +96,9 @@
           The poster image will be used as the main image to represent your film on the marketplace, and, more
           generally, for all film components requiring vertical formats on the marketplace.
         </p>
-        <image-new-uploader [meta]="['movies', 'poster', movieId]" ratio="poster" [form]="form.poster">
+        <image-uploader [meta]="['movies', 'poster', movieId]" ratio="poster" [form]="form.poster">
           <h3 title>Film Poster</h3>
-        </image-new-uploader>
+        </image-uploader>
         <span class="mat-caption">Please note that this document will be available for download on the
           marketplace.</span>
       </article>
@@ -109,9 +109,9 @@
           The banner image will be used as a cover picture on your project page on the marketplace, and, more generally,
           for all film components requiring horizontal formats on the marketplace.
         </p>
-        <image-new-uploader [meta]="['movies', 'banner', movieId]" ratio="banner" [form]="form.banner">
+        <image-uploader [meta]="['movies', 'banner', movieId]" ratio="banner" [form]="form.banner">
           <h3 title>Film Banner Image</h3>
-        </image-new-uploader>
+        </image-uploader>
         <span class="mat-caption">Please note that this document will be available for download on the
           marketplace.</span>
       </article>
