+++ libs/movie/src/lib/movie/form/main/main.module.ts
@@ -15,7 +15,7 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatChipsModule } from '@angular/material/chips';
 
 // Blockframes Media
-import { ImageNewUploaderModule } from '@blockframes/media/image/uploader-new/uploader.module';
+import { ImageUploaderModule } from '@blockframes/media/image/uploader/uploader.module';
 
 // Blockframes UI
 import { ChipsAutocompleteModule } from '@blockframes/ui/static-autocomplete/chips/chips-autocomplete.module';
@@ -44,7 +44,7 @@ import { IsTypePipeModule } from '@blockframes/movie/pipes/is-type.pipe';
     // Other Modules
     TunnelPageModule,
     ChipsAutocompleteModule,
-    ImageNewUploaderModule,
+    ImageUploaderModule,
     StaticSelectModule,
     FormTableModule,
     HasStatusModule,
