+++ libs/movie/src/lib/movie/form/media-images/media-images.module.ts
@@ -18,8 +18,8 @@ import { FormTableModule } from '@blockframes/ui/form/table/form-table.module';
 import { PromotionalLinksModule } from '@blockframes/movie/form/links/promotional-links.module';
 
 // Blockframes Media
-import { FileNewUploaderModule } from '@blockframes/media/file/file-new-uploader/file-uploader.module';
-import { ImageNewUploaderModule } from '@blockframes/media/image/uploader-new/uploader.module';
+import { FileUploaderModule } from '@blockframes/media/file/file-uploader/file-uploader.module';
+import { ImageUploaderModule } from '@blockframes/media/image/uploader/uploader.module';
 
 import { MovieFormMediaImagesComponent } from './media-images.component';
 
@@ -31,9 +31,9 @@ import { MovieFormMediaImagesComponent } from './media-images.component';
     ReactiveFormsModule,
     TunnelPageModule,
     PromotionalLinksModule,
-    FileNewUploaderModule,
+    FileUploaderModule,
     FormTableModule,
-    ImageNewUploaderModule,
+    ImageUploaderModule,
     FlexLayoutModule,
 
     // Material
