+++ libs/movie/src/lib/movie/form/media-notes/notes.component.html
@@ -16,9 +16,9 @@
                 </mat-option>
               </mat-select>
             </mat-form-field>
-            <file-new-uploader test-id="intent-note" [form]="noteForm.get('storagePath')" [meta]="['movies', 'notes', movieId, index]"
+            <file-uploader test-id="intent-note" [form]="noteForm.get('storagePath')" [meta]="['movies', 'notes', movieId, index]"
               accept="pdf">
-            </file-new-uploader>
+            </file-uploader>
             <div fxLayout="column" class="captions">
               <span class="mat-caption">Attach a note / statement of intent from your project's director or
                 producer.</span>
