+++ libs/movie/src/lib/movie/form/media-notes/notes.module.ts
@@ -9,7 +9,7 @@ import { RouterModule } from '@angular/router';
 import { TunnelPageModule } from '@blockframes/ui/tunnel';
 import { FormDisplayNameModule } from '@blockframes/ui/form/display-name/display-name.module';
 import { FormListModule } from '@blockframes/ui/form/list/form-list.module';
-import { FileNewUploaderModule } from '@blockframes/media/file/file-new-uploader/file-uploader.module';
+import { FileUploaderModule } from '@blockframes/media/file/file-uploader/file-uploader.module';
 import { MaxLengthModule } from '@blockframes/utils/pipes';
 import { FileNameModule } from '@blockframes/utils/pipes/fileName.pipe';
 
@@ -30,7 +30,7 @@ import { MovieFormMediaNotesComponent } from './notes.component';
     FormListModule,
     FileNameModule,
     MaxLengthModule,
-    FileNewUploaderModule,
+    FileUploaderModule,
 
     // Material
     MatFormFieldModule,
