+++ libs/movie/src/lib/movie/form/media-videos/media-videos.component.html
@@ -17,10 +17,10 @@
         </mat-list>
       </ng-container>
       <ng-template #noScreener>
-        <file-new-uploader [form]="screenerForm.storagePath" [meta]="['movies', 'screener', movieId]" accept="video">
+        <file-uploader [form]="screenerForm.storagePath" [meta]="['movies', 'screener', movieId]" accept="video">
           filePrivacy="protected">
           <h3 title>Upload Screener</h3>
-        </file-new-uploader>
+        </file-uploader>
         <p class="mat-caption">This video is <strong>private</strong> and <strong>cannot be seen by anyone</strong>
           outside your organization unless you explicitly invite them to a virtual screening.</p>
       </ng-template>
@@ -52,9 +52,9 @@
                 </mat-select>
               </mat-form-field>
             </span>
-            <file-new-uploader [form]="activeForm.get('storagePath')" [meta]="['movies', 'otherVideos', movieId, index]" accept="video">
+            <file-uploader [form]="activeForm.get('storagePath')" [meta]="['movies', 'otherVideos', movieId, index]" accept="video">
               <h3 title>Upload Video</h3>
-            </file-new-uploader>
+            </file-uploader>
 
             <ng-template saveButtonText>Save Video</ng-template>
             <ng-template addButtonText>Add a Video</ng-template>
