+++ libs/movie/src/lib/movie/form/sales-pitch/sales-pitch.component.html
@@ -16,9 +16,9 @@
 
     <section id="sales-video-pitch" fxLayout="column">
       <h3>Sales Video Pitch</h3>
-      <file-new-uploader [form]="salesPitch.get('storagePath')" [meta]="['movies', 'salesPitch', movieId]" accept="video" >
+      <file-uploader [form]="salesPitch.get('storagePath')" [meta]="['movies', 'salesPitch', movieId]" accept="video" >
         <h3 title>Sales Video Pitch</h3>
-      </file-new-uploader>
+      </file-uploader>
     </section>
 
   </ng-container>
