+++ libs/organization/src/lib/organization/forms/organization-form/organization-form.module.ts
@@ -4,7 +4,7 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { OrganizationFormAddressModule } from '../organization-form-address/organization-form-address.module';
 import { OrganizationFormDescriptionModule } from '../organization-form-description/organization-form-description.module';
-import { ImageNewUploaderModule } from '@blockframes/media/image/uploader-new/uploader.module';
+import { ImageUploaderModule } from '@blockframes/media/image/uploader/uploader.module';
 import { FormCountryModule } from '@blockframes/ui/form/country/country.module';
 
 // Material
@@ -22,7 +22,7 @@ import { OrganizationFormComponent } from './organization-form.component';
     CommonModule,
     ReactiveFormsModule,
     FlexLayoutModule,
-    ImageNewUploaderModule,
+    ImageUploaderModule,
     OrganizationFormAddressModule,
     OrganizationFormDescriptionModule,
     FormCountryModule,
