+++ libs/user/src/lib/auth/forms/profile/profile.component.html
@@ -37,8 +37,8 @@
     <mat-card-subtitle>Choose your avatar</mat-card-subtitle>
   </mat-card-header>
   <mat-card-content>
-    <image-new-uploader ratio="square" [meta]="['users', 'avatar', uid]" [form]="form.avatar">
+    <image-uploader ratio="square" [meta]="['users', 'avatar', uid]" [form]="form.avatar">
       <h3 title>Profile Image</h3>
-    </image-new-uploader>
+    </image-uploader>
   </mat-card-content>
 </mat-card>
