+++ libs/user/src/lib/user/user-display.module.ts
@@ -16,7 +16,7 @@ import { MatMenuModule } from '@angular/material/menu';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 
 import { PasswordConfirmModule } from '@blockframes/ui/form/password-confirm/password-confirm.module';
-import { FileNewUploaderModule } from '@blockframes/media/file/file-new-uploader/file-uploader.module';
+import { FileUploaderModule } from '@blockframes/media/file/file-uploader/file-uploader.module';
 import { EditableSidenavModule } from '@blockframes/ui/editable-sidenav/editable-sidenav.module';
 
 export const profileRoutes: Routes = [
@@ -41,7 +41,7 @@ export const profileRoutes: Routes = [
     MatMenuModule,
     EditableSidenavModule,
     PasswordConfirmModule,
-    FileNewUploaderModule,
+    FileUploaderModule,
     RouterModule.forChild(profileRoutes),
   ],
   declarations: [],
